requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-groq
elevenlabs
pypdf
sentence-transformers
faiss-cpu
flask-pymongo
langchain-huggingface